wip : init commit for auth pages.
